Revert "update db seeder and table parameter"
This reverts commit 6b605a7eab96b381cbbcdc6886d46cf0e9b34ae9.

+++ backend-api/database/migrations/2021_12_19_081725_create_schedules_table.php
@@ -27,8 +27,8 @@ class CreateSchedulesTable extends Migration
             $table->time('scheduleTime');
             $table->string('carModel');
             $table->string('carType');
-            $table->float('timeEstimation');
-            $table->float('priceEstimation');
+            $table->integer('timeEstimation');
+            $table->integer('priceEstimation');
             $table->string('scheduleStatus');
             $table->string('description')->nullable();
             $table->timestamps();
